+++ packages/jest-worker/src/workers/NodeThreadsWorker.ts
@@ -182,7 +182,7 @@ export default class ExperimentalWorker
         break;
 
       case PARENT_MESSAGE_CLIENT_ERROR:
-        error = response[4];
+        error = unpackMessage(response[4]);
 
         if (error != null && typeof error === 'object') {
           const extra = error;